Mirnanew

+++ project_3/src/Item/Vinyl.java
@@ -5,7 +5,7 @@ public class Vinyl extends Music{
         super(itemName, itemCondition, purchasePrice, listPrice, salePrice, band, album);
     }
 
-    public Vinyl (String itemName, int itemCondition, Float purchasePrice, Float listPrice, Float salePrice) {
+    public Vinyl (String itemName, int itemCondition, double purchasePrice, double listPrice, double salePrice) {
         super(itemName, itemCondition, purchasePrice, listPrice, salePrice);
     }
 }
